back+front/openai-back/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
azure-common==1.1.28
azure-core==1.35.0
azure-functions==1.23.0
azure-search-documents==11.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
Django==5.2.4
djangorestframework==3.16.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
MarkupSafe==3.0.2
numpy==2.3.1
openai==1.95.1
pandas==2.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
